requirements.txt
tldextract
requests
python-whois
defusedxml
netaddr